+++ src/test/java/uk/q3c/krail/core/option/MockContext0.java
@@ -2,7 +2,7 @@ package uk.q3c.krail.core.option;
 
 import uk.q3c.krail.option.api.Option;
 import uk.q3c.krail.option.test.MockContext;
-import uk.q3c.krail.testutil.option.MockOption;
+import uk.q3c.krail.option.test.MockOption;
 
 import javax.annotation.Nonnull;