+++ src/test/java/uk/q3c/krail/core/option/MockContext2.java
@@ -3,8 +3,8 @@ package uk.q3c.krail.core.option;
 import uk.q3c.krail.option.api.Option;
 import uk.q3c.krail.option.api.OptionKey;
 import uk.q3c.krail.option.test.MockContext;
+import uk.q3c.krail.option.test.MockOption;
 import uk.q3c.krail.testutil.i18n.TestLabelKey;
-import uk.q3c.krail.testutil.option.MockOption;
 
 import javax.annotation.Nonnull;